requirements.txt
annotated-types==0.6.0
pillow==10.3.0
pydantic==2.6.4
pydantic_core==2.16.3
python-dotenv==1.0.1
typing_extensions==4.11.0
